Stage-11-DataBase fix Telegram.kt, structured files
