+++ src/main/kotlin/trainer/model/Question.kt
@@ -1,0 +1,6 @@
+package trainer.model
+
+data class Question(
+    val variants: List<Word>,
+    val correctAnswer: Word,
+)
