+++ src/main/kotlin/trainer/model/Statistics.kt
@@ -1,0 +1,7 @@
+package trainer.model
+
+data class Statistics(
+    val learned: Int,
+    val total: Int,
+    val percentLearned: Int,
+)
